+++ src/js/_init.js
@@ -3,17 +3,23 @@
 
 		// Init Waves
         Waves.init();
-        //Waves.attach('.btn', ['waves-effect', 'waves-light']);
+        Waves.attach('.eq-ui-waves', ['waves-effect']);
+        Waves.attach('.eq-ui-waves-light', ['waves-effect', 'waves-light']);
 
 		// Resize
 		$(window).resize( function() {
 			
 			EqUI.app_bar.update();
 			EqUI.side_nav.update();
+
+			EqUI.collapsible.update();
 			
 			EqUI.site.update();
 
 		});
+		
+		// Show body
+		$('body').css('visibility', 'visible');
 
 	});
 }( jQuery ));